Moves `filters` for Insight to schema package
